#1348: split icmp_embed.rs into module-with-foo layout
Splits the 761-LOC userspace-dp/src/afxdp/icmp_embed.rs into seven
sibling files under icmp_embed/, addressing the two engineering-style
thresholds the file violated:

- try_embedded_icmp_nat_match_from_frame (269 LOC body, Tier-1 >200)
  is now decomposed by outer address family into nat_match_v4.rs
  (102 LOC) and nat_match_v6.rs (123 LOC). The dispatcher in mod.rs
  matches on meta.protocol and routes to the family-specific branch.

- embedded_icmp_return_resolution (10 params, over the 8-param cap)
  collapses to 5 params behind NatMatchCtx<'a> — a single borrow
  bundle of {sessions: &mut SessionTable, forwarding,
  dynamic_neighbors, shared_sessions, shared_nat_sessions,
  shared_forward_wire_sessions}. The ctx is threaded through both
  v4/v6 match fns and reborrowed when handed to return_resolution.

New layout:
  icmp_embed/
    mod.rs              185 LOC — 8 pub(super) wrappers + types
    builders.rs         243 LOC — build_nat_reversed_icmp_error_v4/v6
                                 + finalize_embedded_icmp_resolution
    nat_match_v4.rs     102 LOC — IPv4-outer branch
    nat_match_v6.rs     123 LOC — IPv6-outer branch (preserves the
                                 NPTv6 wire-vs-translated asymmetry)
    parse.rs            149 LOC — EmbeddedV4Header / EmbeddedV6Header
                                 parsers + embedded_reply_key
    return_resolution.rs 31 LOC — embedded_icmp_return_resolution
    session_match.rs     88 LOC — session-only (non-NAT) match path

Each child file is at most 243 LOC; no file is over the Tier-1 200
LOC functional-body threshold (the builders.rs file is two functions
plus finalize; the largest single fn is build_nat_reversed_icmp_error_v4
at ~110 LOC, under the 100-LOC cap by a small margin).

Adversarial review:
- Codex r1 (PLAN-NEEDS-MAJOR): 4 blocking issues — E0364 visibility
  trap, missing mod-declaration change, two-ctx muddiness, NPTv6
  parse safety. All 4 addressed in plan v2.
- Gemini r1: PLAN-READY.
- Codex r2 (sandbox-limited): borrow-checker concern on or_else+map
  chain capturing &mut ctx. Resolved by noting lookup helpers return
  OWNED structs (ResolvedSessionLookup, ForwardSessionMatch) with no
  &SessionTable references — the same .or_else().map() pattern
  compiles cleanly in current code with raw &mut sessions. Build
  confirmed clean. Also flagged file-ambiguity risk (resolved: old
  icmp_embed.rs is deleted in this same commit) and src_wire framing
  (acknowledged as naming/doc, not a Rust newtype).

Design decisions enforced:
- Wrapper-function pattern in mod.rs (NOT pub use ...) avoids E0364
  for pub(super) child items.
- EmbeddedV6Header.src_wire field name + doc keep the NPTv6 wire-vs-
  translated asymmetry visible; translation is applied at the call
  site in nat_match_v6.rs (mirrors icmp_embed.rs:358-360 behavior).
- ONE NatMatchCtx struct — both NAT-match and return-resolution
  paths share the same &mut SessionTable borrow.
- afxdp/mod.rs:78-79 drops '#[path = "icmp_embed.rs"]' and uses
  the bare 'mod icmp_embed;' form, matching the convention used by
  every other directory submodule under afxdp/.
- afxdp/mod.rs:148-154 import block is textually unchanged.

Validation:
- cargo build clean (dev + release).
- cargo test --release: 1506 tests pass (8 embedded_icmp tests
  among them; all pass). The lone failure
  (snat_contract_documents_current_fail_closed_runtime) is a
  pre-existing master-broken doc-guard test unrelated to #1348 —
  verified by failing identically on a clean origin/master clone.
- 5/5 flake pass on embedded_icmp targeted run.
- Go suite: 30 packages clean.
- No new per-packet allocations; NatMatchCtx is a stack-local
  borrow bundle. #[inline] on the dispatch fn keeps the family
  branch foldable.

Closes #1348.
